feat: Refactor order and product models for courier integration
Changed product and product registration weight fields to required decimal type. Added courierId, width, height, length, and weight fields to Order model and database schema. Updated DTOs, service logic, and API client interfaces to use integer dimensions and weights for courier API compatibility. Adjusted seeder and schema files to match new order structure.

+++ BLL/Interfaces/Infrastructure/IGoshipCourierApiClient.cs
@@ -6,7 +6,7 @@ namespace BLL.Interfaces.Infrastructure;
 public interface IGoshipCourierApiClient
 {
     Task<List<RateResponseDTO>> GetRatesAsync(string fromDistrict, string fromCity, string toDistrict,
-        string toCity, decimal cod, decimal amount, decimal width, decimal height, decimal length, decimal weight,
+        string toCity, int cod, int amount, int width, int height, int length, int weight,
         CancellationToken cancellationToken = default);
 
     Task<string> CreateShipmentAsync(string rate, int payer, UserResponseDTO from, UserResponseDTO to,
